+++ backend/CHANGELOG.md
@@ -3,6 +3,10 @@
 ### New
 - Added a separate mode for all the common erc20 paymasters in mind that in future will replace the current erc20 mode
 
+## [3.0.2] - 2025-02-11
+- Using Skandha for gas data for better transaction inclusion.
+- USE_SKANDHA_FOR_GAS_DATA set this property to false to disable.
+
 ## [3.0.1] - 2025-02-04
 ### Fixes
 - Changed error message to more meaningful reply